requirements.txt
PyYAML
pytz
python-telegram-bot
urllib3
requests
requests_html
pdfcrowd
pdf2image
filetype
pymongo[srv]